\name{gicf-package}
\alias{gicf-package}
\docType{package}
\title{
\packageTitle{gicf}
}
\description{
\packageDescription{gicf}
}
\details{
This package optimises the penalised loglikelihood function of a Gaussian covariance graphical model via an iterative coordinate descent algorithm.
}
\author{
\packageAuthor{gicf}

Maintainer: \packageMaintainer{gicf}
}
\references{
Cibinel, L., A. Roverato, and V. Vinciotti (2024). A unified approach to penalized likelihood estimation of covariance matrices in high dimensions. arXiv, arXiv:2410.02403.
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = gicf)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{gicf}}
%\section{Documentation Index}{\packageIndices{gicf}}
