\name{gicf-package}
\alias{gicf-package}
\docType{package}
\title{
\packageTitle{gicf}
}
\description{
\packageDescription{gicf}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{gicf}

Maintainer: \packageMaintainer{gicf}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = gicf)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{gicf}}
%\section{Documentation Index}{\packageIndices{gicf}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
